Removing uin
